[#10834] Add basic PCIe infrastructure.
This is incompatible with previous PCI implementations.
This is, however, generally compatible with the PCI standard.

Previous PCI solutions will be eventually removed, and these interfaces
and classes will be renamed to include PCI explicitly.

This interface is not complete.

At the moment we support double word accesses. More fine grained
accesses can be done via translations on a controller level.