[FIX] Do not stop `SocketServerProvider` when client disconnects
Due to a bug in cancellation tokens handling, disconnection of
a single client caused the whole server to stop. As a result
it was not possible to connect again. This commit fixes the
issue - the server is stopped only when the `Stop` method is
called (directly or as a result of `Dispose`).