[#16176] usbip: Handle USB devices faults
Faulty USB devices might not return
descriptors of expected length, so the
code must be prepared for this.

After the changes, when the server
detects a faulty device it will
print an error to log and return
a default descriptor form to the host
machine (in order not to break the
USB/IP protocol).