[#16176] Rework USB endpoints handling
This commit moves endpoints collection to
`USBDeviceCore` which enables to detect
misconfiguration in case of multi-interface devices.

It also moves setup packet handling to `USBEndpoint`
which allows to use non-zero endpoints as control ones.

It adds `ControlEndpoint` property to `USBDeviceCore`
exposing the default (always-present) EP0.

It also changes the endpoint numbers in USBPendrive
- right now there is EP0 (controll) and two EP1: IN and OUT.