[#16503] TranslationCPU: Fix hiding exceptions
An unhandled exception might cause
the Time Handle to be in a wrong state when
leaving the inner loop of a CPU thread.

This violation of the Time Framework protocol
might lead to another exception
that would hide the original one making
it hard to debug the actual problem.

This commit solves the problem by dropping
the Time Handle deactivation phase when the
exception is detected.