[FIX] Use Marshal.PtrToStringAnsi instead of -Auto
As it turns out, Marshal's PtrToStringAuto method can result in calling
Marshal.PtrToStringUni or Marshal.PtrToStringAnsi version of this function,
based on the "platform".

It's been used in two places for Renode's disassembly logging, which was
broken on Windows. Definitely on Windows the -Auto method results in calling
Unicode version while on Linux and macOS it's the Ansi version.

After changing -Auto to -Ansi method, a proper log is created containing
disassembled instructions on Windows and such change doesn't break logging
neither on Linux nor on macOS.