[#17635] gdb: Fix handling long commands
There was a feature of limiting the length of
packet's `DataAsString` representation to fit it
nicely in the log.

Unfortunatelly it turned out that this property is
also used elsewhere and truncating the string
causes some long GDB commands to fail.

This commit fixes the problem by providing a separate
method for obtaining truncated strings to used in the
log, leaving `DataAsString` as long as it should be.