[#18792] TranslationCPU: Allow to realign mapped memory segments
It's necessary to run platforms with memory segmetns not aligned
to the guest page size.

This feature must be explicitely enabled using `RealignMappedMemorySegments`
property.

NOTE: This doesn't change platform definition,
so memory size reported by the `peripherals` command will
still be unaligned.

Accessing the extended part of a memory segment
from monitor using sysbus `Read/Write` will not be possible.

The CPU read/write instructions however will have access
to the additional memory space (which is inconsistent with the HW).