Fixing CRC calculation and Buffer interaction in K64F Ethernet
The CRC calculation in the Ethernet module was broken. Even if the
application indicated that the CRC needed to be calculated, the Ethernet
simulator did not calculate it.
Also, the internal mechanism of working with RX and TX application ring
buffers of sizes larger than 1 was broken.
The logic took the last and the wrap flag into account, when it only
should have considered the wrap flag.