[#20939] LiteX_I2C: Rework reading data from device
This commit fixes handling of multi-byte reads.
After reading each byte the ACK is sent back to the device
- it requries the clock signal to be toggled.
In the previous implementation this clock change
caused a bit in the Data register to be shifted and lost.

Additionally, the order of bits in the Data register
was wrong as I2C sends the most significant bit first.