[GH#20] ConsoleBackend: Do not crash on unknown console width
In some situations (e.g., on the CI server) `Console.WindowWidth`
is reported as 0 which breaks the math when trying to count lines
for a single message and causes Renode to crash.

In order to avoid that let's verify if we can read a non-zero width
and if not just report repeated lines.