ui: videnanalyzer: add wpf scancode converter
In order to work with raw scancodes on Windows, we need to be able to
convert from Windows scancodes to Renode scancodes.

Add a converter to translate scancodes as necessary.

Note that it has been observed that sometimes, multiple scancodes map to
the same character. In particular, most control codes end up with code
53. Ignore these "translated" codes and just map the raw value.

Signed-off-by: Sean Cross <sean@xobs.io>