ui: videoanalyzer: enable keyboard mapping on windows
This utilises the new WPFToKeyScanCodeConverter in order to map keys on
Windows. This operates independently of the current keymapping, which
e.g. allows for scancodes to be the same regardless of the current
keyboard mapping.

For example, the `Q` key will be reported regardless of whether the
keyboard is in QWERTY or Dvorak mode.

Signed-off-by: Sean Cross <sean@xobs.io>