[#27903] RegisterDescription: Fix regnums for fflags/frm/fcsr
For those three registers we reused rengum of X0 causing
the offsets of registers in g/G commands output to be broken.
GDB counted register X0 as 256 bit long and expected register
a value of register X1 to be located at offset 32 of the g command string.
The rest of registers were expected to be shifted as well.