[FIX] SimpleJson: Properly deserialize integers
Calling `PocoJsonSerializerStrategy.DeserializeObject` with an object
representing a small number (GetType == Int32) resulted in returning
`null`.

As a result, calling `SimpleJson.Deserialize` was also flawed. Calling
`SimpleJson.Deserialize<Dictionary<string,object>>("{\"x\": 2}")`
was returning a dictionary with `null` in an `"x"` key.

Now it's being handled similarly to the double and long values and it
fixes the problem.