[#36075] SensorSamplesFifo: Support feeding samples from binary files
Samples from a binary file are loaded through one or more
`SensorSamplePacket` objects. The `ManagedThread` will then start
feeding samples at the specified time with a frequency specified in the
binary file's header.

Only the delay for the first packet can be specified when loading a
file. Further packets are scheduled based on the time difference between
them and the first packet.