[FIX] SensorSamplesPacket: Extract `Header` struct
`Marshal.SizeOf` and, more importantly, `Marshal.PtrToStructure`
turned out to be unable to work with a structure defined in a generic
class.

It was a problem only on Windows though. There was no such issue under
Mono on Linux and macOS.

Using generic versions of these methods would also solve the problem
but they aren't available in .NET Framework v4.5.