[#37772] MAX32650_RTC: Add `CurrentDateTime` getter
The `BaseDateTime` needs to be known to be able to translate RTC
counters to a DateTime. The RTC itself only increments counters without
any definition of what "real" date and time they indicate.

It will also be crucial for setting RTC counters from the given
DateTime.

The start of Unix Epoch is the default base but it can be modified with
a constructor parameter.