[#46582] NativeBinder: Allow for optional imports
This commit adds the `Optional` option to `ImportAttribute`.
Symbols with this option set can be missing in imported library
and it will not throw an exception on import. This Exception will
be delayed to the first call of undefined method.

It allows for easier backwards compatibility, new native methods can
be marked `Optional` and used when needed.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>