[#48754] PythonEngine: Remove `machine` and `sysbus` variables
Initializing those variables resulted in temporary pausing of the machine
(as a result of the TryGetMachineForPeripheral->RegisterBusController->ObtainPausedState call chain),
which in turn could interfere with the GDB integration layer - as the pause event would be
reported to the external GDB instance using the SIGTERM signal and would put it out of sync with the simulation.

To reference `machine` and `systembus` from Python use`self.GetMachine()` and `self.GetMachine().SystemBus` respectively.