[#50236] MiV_CoreGPIO: Add OUTBUFF register
In order to run tests using the MiV_CoreGPIO peripheral it is required
that the OUTBUFF tag has both read and write functionality. Currently if
only has the write functionality as is using a tag. In order to allow
for both read and write functionality it is necessary to change this tag
to a flag.

Signed-off-by: Eoin Dickson <Eoin.dickson@microchip.com>