Fixed issues with DDR Training
This module needs to supply more expected return values for training
to complete successfully.

Also took the opportunity to add a 'signature' register, which the
software can use to choose to skip training. Training has no impact
in Renode and takes several minutes to complete - so an option to skip
it is merited.

Signed-off-by: Edward Connell <Edward.Connell@microchip.com>