[#54450] GaislerEth: Remove HostToNetworkOrder
This was previously needed to use this peripheral on big-endian
platforms. What would happen is that the values read from memory through
the system bus would be byte-swapped, so we needed to swap them again in
the peripheral (the fact that we use htonl to do it is a red herring).
Now this is not needed anymore.