[#53750] SystemBus: Add context to UnmapMemory
The changes in AddMappings make no functional difference but the code is
more readable this way.

`UnmapMemory` uses `ICPU context` because it's tricky. `context == null`
is used as "for all CPUs" so with `context: cpu as ICPUWithMappedMemory`
it wouldn't be possible to distinguish whether `context` was null or if
the CPU passed doesn't implement `ICPUWithMappedMemory`.