[#57659] TranslationCPU: Add hook to obtain cpuId
In SMP, cpuId might be different than the slot (registration id). This
Id will be used as cpuIndex, when identifying CPUs for SMP (e.g. on Arm
this is the ID that can be seen in MPIDR register). Slot is used to
identify the CPU from the perspecitve of the framework itself (when
identifying CPU within peripherals' code).