Evaluated value of CancellationToken before initiating timer
If the token is evaluated after the timer checks it then the cancellation
token source has changed, and the new value is checked. This caused
every timer to cause an idle line interrupt. Evaluating the token before
starting the timer means that the cancellation status is not overwritten
by new tokens.