STM32 UART/DMA Transfer Request Fixes
This commit attempts to reproduce the behaviour descrbed in Figure
316. "Reception using DMA" of the STM32F4 Referance Manual (RM0090)

We implement the IDMA interface in the DMA, and from the UART we
request a DMA transfer when value enters data register.
If there is a dma peripheral assigned, then when a value appears in the
DR, request a transfer.

When a transfer is requested by a peripheral a counter keeps track of
how many have been requested. The counter is decremented each time a
transfer happens. When the DMA is re-enabled, any transfers that were
requested while it was disabled are processed. This is useful for
multi-buffering UART, eg with the Zephyr asynchronous UART API.