TranslationCPU: Always eventually handle IRQs
If an interrupt is triggered while the CPU is stopped, then we don't do
anything about it. However, this is the only chance we have to handle
the rising edge of the incoming IRQ. If the interrupt controller tries
to trigger a new interrupt after the CPU has been resumed, the GPIO will
remember that it was already set and won't call OnGPIO. This will cause
all further IRQs to be ignored.

Fix this by enqueuing deferred IRQs until the CPU is resumed (or one of
the other conditions changes). This ensures that we eventually handle
all calls to OnGPIO.

Fixes: 160f4d18 ("Initial commit of the Emul8 framework.")