[#78961] tlib: Implement Hash Table-Based Store-Test (HST) for atomics
This is a new scheme that fixes RISC-V LR/SC invalidation and replaces
the global memory lock in RISC-V with more fine-grained locking around
memory regions.

The same changes can also be made to eliminate the global memory lock
from other guest platforms. This will be done as a later step.